Use structs for body parts
The go code assembles the unicorn piece by piece by appending to a
collection and has some special code to walk the collection.  My rust
skills are insufficient to duplicate this.

Instead we now have various body parts that get assembled to build a
unicorn body.  First the face, then the neck, then the torso, (we
already had legs).

This directly walks the unicorn's body for the various rotations, but
some items are walked twice (which the go code does not do through the
special code).  Instead we can special-case the attachment points to
only a connection point once.

Unfortunately this means the structures must have knowledge of each
other's implementations.  Until I can figure out how to duplicate the
special code this will have to do.